Issue 90
consistent search across multi panel and single panel. single points will be selected instead of multiple points when crosshair is taken over the chart in multi panel. the issue was with the tolerance that allowed multiple points. this was replaced with binary search closest which returns only the single closest point