issue 89, issue 94, issue 84(duplicate)
destroy event and null pointer issue resolved. state of legend now passes from single to multi and multi to single(if legend off in single, it will be off if multi is toggled on and vice versa). Nick will implement this logic for the remaining toggles by following the logic I have written.